часть новой сортировки

+++ src/main/java/ru/yalabuniversity/homework/lecture3/filesort/Sorter.java
@@ -7,83 +7,33 @@ import java.nio.file.StandardCopyOption;
 import java.util.Arrays;
 
 public class Sorter {
-    private int fileSize;
 
-    public Sorter(int fileSize) {
-        this.fileSize = fileSize;
-    }
-
-    public File sortFile(File dataFile) throws IOException {
-        if(fileSize <= 0) {
-            throw new IOException("Введите количество строк больше 0");
+    public File sortFile(int fileSize, File dataFile) throws IOException {
+        if(fileSize < 0) {
+            throw new IOException("Вы введли отрицательное число");
         }
         BufferedReader bufferedReader = null;
         BufferedWriter bufferedWriter = null;
         int middle = 0;
-        try {
-            int size = 2;
-            int left = 0;
-            int right = size - 1;
-            while (size <= fileSize) {
-                File targetFile = new File(".\\src\\main\\resources\\files\\outputData.txt");
-                bufferedReader = new BufferedReader(new FileReader(dataFile));
-                bufferedWriter = new BufferedWriter(new FileWriter(targetFile));
-                while (right <= fileSize) {
-                    int k = 0;
-                    long[] partArr = new long[size];
-                    while (k < size) {
-                        partArr[k] = Long.parseLong(bufferedReader.readLine());
-                        k++;
-                    }
-                    long[] arrayA = new long[size / 2];
-                    System.arraycopy(partArr, 0, arrayA, 0, size / 2);
-                    Arrays.sort(arrayA);
-                    long[] arrayB = new long[size - size / 2];
-                    System.arraycopy(partArr, size / 2, arrayB, 0, size - size / 2);
-                    Arrays.sort(arrayB);
-                    long[] resultArr = mergeArray(arrayA, arrayB);
-                    int r = 0;
-                    while (r < resultArr.length) {
-                        bufferedWriter.write(resultArr[r] + "\n");
-                        r++;
-                    }
-                    bufferedWriter.flush();
-                    left += size;
-                    right = left + size;
-                }
-                if (right > fileSize) {
-                    right -= size;
-                    middle = right;
-                    while (right < fileSize) {
-                        bufferedWriter.write(bufferedReader.readLine() + "\n");
-                        right++;
-                    }
-                    bufferedWriter.flush();
-                }
-                size *= 2;
-                bufferedReader.close();
-                bufferedWriter.close();
-                Files.copy(Paths.get(targetFile.toURI()), Paths.get(dataFile.toURI()), StandardCopyOption.REPLACE_EXISTING);
-                right = size - 1;
-                left = 0;
-                if (!targetFile.delete()) {
-                    throw new IOException("Возникла ошибка при удалении");
-                }
-            }
-            if (middle != fileSize) {
-                File targetFile = new File(".\\src\\main\\resources\\files\\outputData.txt");
-                bufferedReader = new BufferedReader(new FileReader(dataFile));
-                bufferedWriter = new BufferedWriter(new FileWriter(targetFile));
+        int size = 2;
+        int left = 0;
+        int right = size - 1;
+        while (size <= fileSize) {
+            File targetFile = new File(".\\src\\main\\resources\\files\\outputData.txt");
+            bufferedReader = new BufferedReader(new FileReader(dataFile));
+            bufferedWriter = new BufferedWriter(new FileWriter(targetFile));
+            while (right <= fileSize) {
                 int k = 0;
-                long[] partArr = new long[fileSize];
-                while (k < fileSize) {
+                long[] partArr = new long[size];
+                while (k < size) {
                     partArr[k] = Long.parseLong(bufferedReader.readLine());
                     k++;
                 }
-                long[] arrayA = new long[middle];
-                System.arraycopy(partArr, 0, arrayA, 0, middle);
-                long[] arrayB = new long[fileSize - middle];
-                System.arraycopy(partArr, middle, arrayB, 0, fileSize - middle);
+                long[] arrayA = new long[size / 2];
+                System.arraycopy(partArr, 0, arrayA, 0, size / 2);
+                Arrays.sort(arrayA);
+                long[] arrayB = new long[size - size / 2];
+                System.arraycopy(partArr, size / 2, arrayB, 0, size - size / 2);
                 Arrays.sort(arrayB);
                 long[] resultArr = mergeArray(arrayA, arrayB);
                 int r = 0;
@@ -92,17 +42,59 @@ public class Sorter {
                     r++;
                 }
                 bufferedWriter.flush();
-                bufferedReader.close();
-                bufferedWriter.close();
-                Files.copy(Paths.get(targetFile.toURI()), Paths.get(dataFile.toURI()), StandardCopyOption.REPLACE_EXISTING);
-                if (!targetFile.delete()) {
-                    throw new IOException("Возникла ошибка при удалении");
-                }
+                left += size;
+                right = left + size;
             }
-        } finally {
+            if (right > fileSize) {
+                right -= size;
+                middle = right;
+                while (right < fileSize) {
+                    bufferedWriter.write(bufferedReader.readLine() + "\n");
+                    right++;
+                }
+                bufferedWriter.flush();
+            }
+            size *= 2;
             bufferedReader.close();
             bufferedWriter.close();
+            Files.copy(Paths.get(targetFile.toURI()), Paths.get(dataFile.toURI()), StandardCopyOption.REPLACE_EXISTING);
+            right = size - 1;
+            left = 0;
+            if (!targetFile.delete()) {
+                throw new IOException("Возникла ошибка при удалении");
+            }
         }
+        if (middle != fileSize) {
+            File targetFile = new File(".\\src\\main\\resources\\files\\outputData.txt");
+            bufferedReader = new BufferedReader(new FileReader(dataFile));
+            bufferedWriter = new BufferedWriter(new FileWriter(targetFile));
+            int k = 0;
+            long[] partArr = new long[fileSize];
+            while (k < fileSize) {
+                partArr[k] = Long.parseLong(bufferedReader.readLine());
+                k++;
+            }
+            long[] arrayA = new long[middle];
+            System.arraycopy(partArr, 0, arrayA, 0, middle);
+            long[] arrayB = new long[fileSize - middle];
+            System.arraycopy(partArr, middle, arrayB, 0, fileSize - middle);
+            Arrays.sort(arrayB);
+            long[] resultArr = mergeArray(arrayA, arrayB);
+            int r = 0;
+            while (r < resultArr.length) {
+                bufferedWriter.write(resultArr[r] + "\n");
+                r++;
+            }
+            bufferedWriter.flush();
+            bufferedReader.close();
+            bufferedWriter.close();
+            Files.copy(Paths.get(targetFile.toURI()), Paths.get(dataFile.toURI()), StandardCopyOption.REPLACE_EXISTING);
+            if (!targetFile.delete()) {
+                throw new IOException("Возникла ошибка при удалении");
+            }
+        }
+        bufferedReader.close();
+        bufferedWriter.close();
         return dataFile;
     }
